requirements.txt
matplotlib 
scikit-image 
scikit-learn 
scipy 
numpy 
torch 
pytorch==1.2.0
torchvision==0.4.0
pillow==5.2.0



